==> 1-data/1-notebooks/02-census-acs.Rmd <==
---
df_print: tibble
output:
  html_notebook:
    default
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

```{r low-inc-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(ggiraph)
library(leaflet.extras)
library(viridisLite)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()
kc_tr_acs <- acs::geo.make(state = "WA",county = "King", tract = "*")

```

### ACS Geo Set
This geoset is mixture of geography type including:

  - Seattle (place)
  - All tracts in/partially within Seattle
  - Tract-neighborhoods (combined tract groups)
  - Tract-neighborhood groups (combined tract groups)
  
```{r census-acs, echo = TRUE}

# Setup the ACS geographies: Seattle (tracts), YCC (tracts), Seattle (place)
# -----------------------------------------------------------------------------

if(!exists('ycc_geo_acs')){
        ycc_all_sf <- 
        read_rds(root_file('1-data/4-interim/ycc-all-sf.rds')) %>% 
        mutate(NAME_FCT = factor(NAME_FULL,
                                 levels = c('Capitol Hill',
                                            'Central Area',
                                            'First Hill',
                                            'Chinatown International District'),
                                 ordered = TRUE)) %>% 
        select(everything(),geom)

get_geoid <- function(x,nm){
        x %>% filter(NAME %in% nm & !is.na(GEOID6)) %>% select(GEOID6) %>% unlist(use.names = FALSE) %>% as.numeric
}

tr_ch <- get_geoid(ycc_all_sf,'CH')
tr_ca <- get_geoid(ycc_all_sf,'CA')
tr_fh <- get_geoid(ycc_all_sf,'FH')
tr_cid <- get_geoid(ycc_all_sf,'CID')
tr_ycc <- get_geoid(ycc_all_sf,c('CH','CA','FH','CID'))
tr_sea <- 
        ycc_all_sf %>% 
        filter(!is.na(GEOID6)) %>% 
        transmute(GEOID6 = as.numeric(GEOID6)) %>% unlist(use.names = FALSE)
#
# Final Object
ycc_geo_acs <- c(
        acs::geo.make(state = "WA",county = "King", tract = tr_sea),
        acs::geo.make(state = "WA",county = "King", tract = tr_ch, combine = TRUE, combine.term = 'CH'),
        acs::geo.make(state = "WA",county = "King", tract =tr_ca, combine = TRUE, combine.term = 'CA'),
        acs::geo.make(state = "WA",county = "King", tract = tr_fh, combine = TRUE, combine.term = 'FH'),
        acs::geo.make(state = "WA",county = "King", tract = tr_cid, combine = TRUE, combine.term = 'CID'),
        acs::geo.make(state = "WA",county = "King", tract = tr_ycc, combine = TRUE, combine.term = 'YCC'),
        acs::geo.make(state = 'WA', place = "Seattle")
)



ycc_geo_acs %>% write_rds(root_file('1-data/4-interim/ycc-geo-acs.rds'))

}


```
