==> 1-data/1-notebooks/04-combined.Rmd <==
---
df_print: tibble
output:
  html_notebook:
    code_folding: hide
  pdf_document:
    keep_tex: yes
always_allow_html: yes
---

```{r comb-setup, echo = FALSE, warning=FALSE,message=FALSE,comment=FALSE}
library(plyr)
library(knitr)
library(rprojroot)
library(rgdal)
library(sp)
library(rgeos)
library(tigris)
library(leaflet)
library(ggthemes)
library(magrittr)
library(stringr)
library(downloader)
library(webshot)
library(htmltools)
library(gplots)
library(ggmap)
library(shiny)
library(htmlwidgets)
library(readxl)
library(acs)
library(RColorBrewer)
library(tidyverse)
library(miscgis)
library(operator.tools)
library(ggiraph)
library(leaflet.extras)
library(viridisLite)
library(sf)
root <- rprojroot::is_rstudio_project
root_file <- root$make_fix_file()
opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE, comment=FALSE)
green <- miscgis::miscgis_pals$tableau_cat[["green"]]
blue <- miscgis::miscgis_pals$tableau_cat[["blue"]]
orange <- miscgis::miscgis_pals$tableau_cat[["orange"]]
red <- miscgis::miscgis_pals$tableau_cat[["red"]]
teal <- miscgis::miscgis_pals$tableau_cat[["teal"]]
pal_rgb_4 <- miscgis_pals$tableau_cat[c("red","gold","green","blue")] %>% unlist %>% palette()

```
### Combined Census Indicators

```{r comb-make-combined-sf}

combined_sf <-
        list.files(root_file('1-data/4-interim'),pattern = 'sf\\.rds$') %>%
        .[!str_detect(.,'ycc-all-sf\\.rds')] %>% 
        paste0(root_file('1-data/4-interim/'),.) %>% 
        map(read_rds) %>% 
        do.call('cbind',.) %>% 
        select(JOIN_ID,
               matches('NAME'),
               matches('GEO'),
               everything(),
               -matches('\\.[[:digit:]]'))

combined_sf %>% write_rds(root_file('1-data/5-tidy/combined-sf.rds'))
        

```

### "Long" Version

```{r comb-make-long}

first_not_na <- function(x){
        
        x[!is.na(x)][1]
        
        }

replace_empty <- function(x){
        ifelse(nchar(x)==0,NA_character_,x)
}

comb_long_sf <- 
        combined_sf %>%
        mutate_if(is.factor,as.character) %>% 
        gather(VAR,PCT,matches('EST|MOE|UPPER|LOWER')) %>% 
        mutate(Variable = case_when(str_detect(.$VAR,'WORK_POOR') ~ 'Working_Poor',
                                    str_detect(.$VAR,'POC') ~ 'People_of_Color',
                                    str_detect(.$VAR,'HOUS_BURDEN') ~ 'Housing_Cost_Burdened',
                                    str_detect(.$VAR,'LMTD_ENG') ~ 'Limited_English',
                                    str_detect(.$VAR,'FRN_BORN') ~ 'Foreign_Born',
                                    str_detect(.$VAR,'UNEMPLOYED') ~ 'Unemployment'),
               PCT_TYPE = case_when(str_detect(.$VAR,'EST') ~ 'Estimate',
                                    str_detect(.$VAR,'MOE') ~ 'Margin_of_Error',
                                    str_detect(.$VAR,'UPPER') ~ 'Upper_Confint',
                                    str_detect(.$VAR,'LOWER') ~ 'Lower_Confint')) %>% 
        spread(PCT_TYPE,PCT) %>% 
        group_by(JOIN_ID,Variable) %>% 
        summarise_all(first_not_na) %>%
        as.data.frame() %>%
        mutate_at(matches('NAME'),replace_empty) %>% 
        mutate(geom = st_sfc(geom)) %>% 
        st_sf() %>% 
        miscgis::coerce_to_geom(st_multipolygon) %>% 
        select(JOIN_ID,
               Variable,
               Estimate,
               Margin_of_Error,
               Upper_Confint,
               Lower_Confint,
               everything(),
               -VAR) %>% 
        mutate(NAME_FCT = factor(NAME,
                                 levels = c('CH',
                                            'CA',
                                            'FH',
                                            'CID',
                                            'YCC',
                                            'SEA'),
                                 ordered = TRUE),
               NAME_FULL_FCT = factor(NAME_FULL,
                                      levels = c('Capitol Hill',
                                                 'Central Area',
                                                 'First Hill',
                                                 'Chinatown International District',
                                                 'YCC Neighborhoods',
                                                 'Seattle'),
                                      ordered = TRUE),
               NAME_FULL_FCT_NHOODS = factor(NAME_FULL,
                                             levels = c('Capitol Hill',
                                                        'Central Area',
                                                        'First Hill',
                                                        'Chinatown International District'),
                                             ordered = TRUE),
               Variable = str_replace_all(Variable,'_'," "))

comb_long_sf %>% write_rds(root_file('1-data/5-tidy/comb-long-sf.rds'))


```
